refactor : PlayingGame -> RacingGame 이름 수정
